test: simplify healthcheck regression tests
